requirements.txt
largestinteriorrectangle==0.2.1
llvmlite==0.43.0
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
pygame-ce==2.5.2
pyserial==3.5
PyYAML==6.0.2
stitching==0.6.1
